Use bigint instead of int8

+++ baremaps-core/src/main/java/org/apache/baremaps/openstreetmap/postgres/PostgresNodeRepository.java
@@ -257,7 +257,7 @@ public class PostgresNodeRepository implements NodeRepository {
     }
     try (Connection connection = dataSource.getConnection();
         PreparedStatement statement = connection.prepareStatement(deleteIn)) {
-      statement.setArray(1, connection.createArrayOf("int8", keys.toArray()));
+      statement.setArray(1, connection.createArrayOf("bigint", keys.toArray()));
       statement.execute();
     } catch (SQLException e) {
       throw new RepositoryException(e);
